Additional cleanup.

+++ v5.1.2/Unit/src/unit/TestNetworkBinaryCodec.cs
@@ -1,50 +1,6 @@
-// This source code is dual-licensed under the Apache License, version
-// 2.0, and the Mozilla Public License, version 1.1.
-//
-// The APL v2.0:
-//
-//---------------------------------------------------------------------------
-//   Copyright (c) 2007-2016 Pivotal Software, Inc.
-//
-//   Licensed under the Apache License, Version 2.0 (the "License");
-//   you may not use this file except in compliance with the License.
-//   You may obtain a copy of the License at
-//
-//       https://www.apache.org/licenses/LICENSE-2.0
-//
-//   Unless required by applicable law or agreed to in writing, software
-//   distributed under the License is distributed on an "AS IS" BASIS,
-//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//   See the License for the specific language governing permissions and
-//   limitations under the License.
-//---------------------------------------------------------------------------
-//
-// The MPL v1.1:
-//
-//---------------------------------------------------------------------------
-//  The contents of this file are subject to the Mozilla Public License
-//  Version 1.1 (the "License"); you may not use this file except in
-//  compliance with the License. You may obtain a copy of the License
-//  at http://www.mozilla.org/MPL/
-//
-//  Software distributed under the License is distributed on an "AS IS"
-//  basis, WITHOUT WARRANTY OF ANY KIND, either express or implied. See
-//  the License for the specific language governing rights and
-//  limitations under the License.
-//
-//  The Original Code is RabbitMQ.
-//
-//  The Initial Developer of the Original Code is Pivotal Software, Inc.
-//  Copyright (c) 2007-2016 Pivotal Software, Inc.  All rights reserved.
-//---------------------------------------------------------------------------
-
 using NUnit.Framework;
-
-using System.IO;
-
-using RabbitMQ.Client;
-using RabbitMQ.Client.Impl;
 using RabbitMQ.Util;
+using System.IO;
 
 namespace RabbitMQ.Client.Unit
 {
@@ -112,25 +68,24 @@ namespace RabbitMQ.Client.Unit
             Assert.AreEqual(0x89AB, Reader(new byte[] { 0x89, 0xAB }).ReadUInt16());
         }
 
-
         [Test]
         public void TestWriteInt32_positive()
         {
-            m_w.Write((int)0x12345678);
+            m_w.Write(0x12345678);
             Check(m_w, new byte[] { 0x12, 0x34, 0x56, 0x78 });
         }
 
         [Test]
         public void TestWriteInt32_negative()
         {
-            m_w.Write((int)-0x12345678);
+            m_w.Write(-0x12345678);
             Check(m_w, new byte[] { 0xED, 0xCB, 0xA9, 0x88 });
         }
 
         [Test]
         public void TestWriteUInt32()
         {
-            m_w.Write((uint)0x89ABCDEF);
+            m_w.Write(0x89ABCDEF);
             Check(m_w, new byte[] { 0x89, 0xAB, 0xCD, 0xEF });
         }
 
@@ -146,25 +101,24 @@ namespace RabbitMQ.Client.Unit
             Assert.AreEqual(0x89ABCDEF, Reader(new byte[] { 0x89, 0xAB, 0xCD, 0xEF }).ReadUInt32());
         }
 
-
         [Test]
         public void TestWriteInt64_positive()
         {
-            m_w.Write((long)0x123456789ABCDEF0);
+            m_w.Write(0x123456789ABCDEF0);
             Check(m_w, new byte[] { 0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0xF0 });
         }
 
         [Test]
         public void TestWriteInt64_negative()
         {
-            m_w.Write((long)-0x123456789ABCDEF0);
+            m_w.Write(-0x123456789ABCDEF0);
             Check(m_w, new byte[] { 0xED, 0xCB, 0xA9, 0x87, 0x65, 0x43, 0x21, 0x10 });
         }
 
         [Test]
         public void TestWriteUInt64()
         {
-            m_w.Write((ulong)0x89ABCDEF01234567);
+            m_w.Write(0x89ABCDEF01234567);
             Check(m_w, new byte[] { 0x89, 0xAB, 0xCD, 0xEF, 0x01, 0x23, 0x45, 0x67 });
         }
